requirements.txt
djangorestframework-simplejwt==5.3.1
pytest==8.3.3
pytest-django==4.9.0
pytest-factoryboy==2.7.0
django-ratelimit==4.1.0